# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ct-project-api",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://ct-project.pp.ua/api/"
        },
        {
            "url": "http://172.19.0.1/api/"
        }
    ],
    "paths": {
...
        "/v1/resend-verification-email": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "6c8288970086eaedc92106f45c029648",
                "responses": {
                    "200": {
                        "description": "Send verification email data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EmailVerification"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "If email is verified already",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Your email is already verified."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/email-verify/{id}": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "c4a82e83c15136210a0fe0a907996d8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "expires",
                        "in": "query",
                        "description": "timestamp",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "query",
                        "description": "Hashed email",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "signature",
                        "in": "query",
                        "description": "electronic signature",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verify your email or email is already verified",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Your email has been verified."
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "If url expired or invalid",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Invalid/Expired url provided."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/reset-password": {
            "post": {
                "tags": [
                    "Password"
                ],
                "operationId": "17e2d5e32f73e63e4d6d10b61dae2088",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Only emails of existing users, max:255 symbols",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generates reset password token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ResetPassword"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "data": {
                                        "email": "example@gmail.com",
                                        "token": "736b32a5265d825732add264f4f85d9e090e4231193a2ad916d962b5a9aca089"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "The user's email is not exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Error"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "data": {
                                        "message": "The user is not exist.",
                                        "code": 400
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/change-password/{token}": {
            "post": {
                "tags": [
                    "Password"
                ],
                "operationId": "2687b766a372c911895c6c28ee27ca3c",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The password has been changed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "message": "The password has been changed."
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid token or email",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Error"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "data": {
                                        "message": "Invalid token or email.",
                                        "code": 400
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error": {
                "description": "Show error exception",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "ValidationErrors": {
                "description": "Validation Errors",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "description": "each key describes error message",
                        "properties": {
                            "query or path name": {
                                "type": "array",
                                "items": {}
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "example": {
                    "message": "The email field must be a valid email address. (and 3 more errors)",
                    "errors": {
                        "email": [
                            "The email field must be a valid email address."
                        ],
                        "password": [
                            "The password field confirmation does not match.",
                            "The password field must contain at least one uppercase and one lowercase letter.",
                            "The password field must contain at least one number."
                        ]
                    }
                }
            },
...
            "ResetPassword": {
                "description": "Show email and generated reset token",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "EmailVerification": {
                "description": "Show email verification data",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "expires": {
                        "type": "integer"
                    },
                    "hash": {
                        "type": "string"
                    },
                    "signature": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}